% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopAudio.R
\name{chopAudio}
\alias{chopAudio}
\title{Chop audio file into several segments}
\usage{
chopAudio(file, timethrs, plot = F, saveWav = T, ...)
}
\arguments{
\item{file}{Wave file address.}

\item{timethrs}{The time constant used to segment the audio.}

\item{...}{Arguments passed to seewave::timer.}
}
\description{
Chop audio file into several segments
}
\details{
Chops an audio file into several segments defined by a time 
constant.

The script uses the seewave::timer function to ensure that the segment
boundaries do not cut across sounds.
}
\keyword{audio,segmentation}
