% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreHeatmap.R
\name{exploreHeatmap}
\alias{exploreHeatmap}
\title{Plot the spectrogram of a representative sound in a similariy heatmap}
\usage{
exploreHeatmap(coord, pos, score, sound2files, sound1file, title)
}
\arguments{
\item{coord}{Coordinates in the heatmap to explore.}

\item{sound2files}{Vector of file addresses of song 2 for each score value}

\item{sound1file}{File address of song 1.
used to produce the headmap.}
}
\description{
Plot the spectrogram of a representative sound in a similariy heatmap
}
