% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceImitation.R
\name{referenceImitation}
\alias{referenceImitation}
\title{Evaluate how well a reference song is imitated}
\usage{
referenceImitation(pos, score, plot = F, na.rm = F, valid = NULL,
  smooth = F)
}
\arguments{
\item{plot}{Should interpolated scores be plotted?}

\item{na.rm}{Should missing columns be removed from reference song?}
}
\description{
Evaluate how well a reference song is imitated
}
\details{
Missing columns in the reference song are normally caused by silence.
}
