% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAPfeatures.R
\name{SAPfeatures}
\alias{SAPfeatures}
\title{Extract spectral features}
\usage{
SAPfeatures(wave, fmin = 0, fmax = NULL, wl, ovlp)
}
\arguments{
\item{fmin}{Minimum frequency at which to analyze sound.}

\item{fmax}{Maximum frequency at which to analyze sound.}

\item{wl}{Window size for the analysis.}

\item{ovlp}{Overlap between two sucessive windows (in %).}

\item{sound}{Wave object.}
}
\description{
The features that are extracted are similar or equivalent to those
used in Sound Analysis Pro 2011 for the calculation of similarity indexes.
When possible, functions from `seewave` were using.
}
