% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityDistr.R
\name{similarityDistr}
\alias{similarityDistr}
\title{Plots the distribution of matches to a reference song}
\usage{
similarityDistr(pos, score, title = NULL, plot = T, ylab = NULL,
  plottype = "point", alpha = NULL)
}
\arguments{
\item{pos}{Vector of positions in the reference song that have been matched.}

\item{score}{Vector of similarity scores.}

\item{title}{Title used for the plot.}

\item{plot}{Should the plot be printed?}

\item{ylab}{Title for the y-axis.}

\item{plottype}{Typle of geom to use for the plot}

\item{alpha}{Value of alpha argument in `ggplot2::geom_point()`, if using
`plottype` "point".}
}
\description{
Plots the distribution of matches to a reference song
}
\details{
Each element in `score` must correspond to an element in `pos`.

  If geom point is used for the plot, specify a value for the alpha argument.
}
