% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoHeatmap.R
\name{similarityHeatmap}
\alias{similarityHeatmap}
\title{Generate a heatmap of matches to a reference song}
\usage{
similarityHeatmap(pos, score, title = NULL, plot = T)
}
\arguments{
\item{pos}{Vector of positions in the reference song that have been matched.}

\item{score}{Vector of similarity scores.}

\item{title}{Title used for the plot.}

\item{plot}{Should the plot be printed?}
}
\description{
Generate a heatmap of matches to a reference song
}
\details{
Each element in `score` must correspond to an element in `pos`.
}
