% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAPfeatures.R
\name{SAPfeatures}
\alias{SAPfeatures}
\title{Extract spectral features}
\usage{
SAPfeatures(wave, fmin = 0, fmax = NULL, wl, ovlp, threshold = 5,
  segment = T, plotsegments = F, export_spectro = F)
}
\arguments{
\item{fmin}{Minimum frequency (in Hz) from which to analyze sound.}

\item{fmax}{Maximum frequency (in Hz) up to which to analyze sound.}

\item{wl}{Window size for the analysis.}

\item{ovlp}{Overlap between two sucessive windows (in %).}

\item{sound}{Wave object.}
}
\description{
The features that are extracted are similar or equivalent to those
used in Sound Analysis Pro 2011 for the calculation of similarity indexes.
When possible, functions from `seewave` were used. The calculation of the
pitch is not yet similar to that in SAP.
}
\details{
The output also includes dominant frequency, which is not generally used
in SAP.
}
